source/requirements.txt
diffusers==0.12.1
einops==0.6.0
k_diffusion==0.0.12
lark==1.1.5
numpy==1.24.1
Pillow==9.4.0
safetensors==0.2.8
tqdm==4.64.1
transformers==4.26.0
simple_websocket_server==0.4.2
websocket_client==1.5.0
bson==0.5.10
Send2Trash==1.8.0
PyOpenGL==3.1.6
PyQt5==5.15.7
psutil==5.9.2